Initial implementation of BoardPage
A `BoardPage` is an interactive widget that shows the player a
minesweeper game board. This is an initial implementation, so you can't
still customize a game's board (e.g., setting number of rows, columns
and mines). That'll be done on another change.

This change also doesn't come with unit tests. That's also coming in a
later commit.